refactor: change getUserSession to Services
